+++ lib/karafka/pro/processing/strategies/lrj_mom.rb
@@ -60,6 +60,11 @@ module Karafka
             coordinator.on_revoked do
               coordinator.revoke
             end
+
+            Karafka.monitor.instrument('consumer.revoke', caller: self)
+            Karafka.monitor.instrument('consumer.revoked', caller: self) do
+              revoked
+            end
           end
         end
       end